requirements.txt
Django
django-jazzmin 
django-ninja
djangorestframework 
django-import-export
